re-rorged `android` package, added `launcher` icon test
