+++ testdata/roller-skate.png
